+++ tokio/src/lib.rs
@@ -479,7 +479,6 @@ pub mod io;
 pub mod net;
 
 mod loom;
-mod park;
 
 cfg_process! {
     pub mod process;
