+++ tokio/src/runtime/mod.rs
@@ -181,6 +181,8 @@ pub(crate) mod context;
 
 pub(crate) mod coop;
 
+pub(crate) mod park;
+
 mod driver;
 
 pub(crate) mod scheduler;
